Refactor common test functions into utils
