+++ test/e2e/reconciler/cluster/controller_test.go
@@ -21,34 +21,21 @@ import (
 	"embed"
 	"errors"
 	"fmt"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/google/go-cmp/cmp"
 
-	corev1 "k8s.io/api/core/v1"
-	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	apiextensionsv1client "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
-	kerrors "k8s.io/apimachinery/pkg/util/errors"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/watch"
-	kubernetesclientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
-	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/kcp-dev/kcp/config"
-	clusterv1alpha1 "github.com/kcp-dev/kcp/pkg/apis/cluster/v1alpha1"
-	kcpclientset "github.com/kcp-dev/kcp/pkg/client/clientset/versioned"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
 	"github.com/kcp-dev/kcp/test/e2e/reconciler/cluster/apis/wildwest"
 	wildwestv1alpha1 "github.com/kcp-dev/kcp/test/e2e/reconciler/cluster/apis/wildwest/v1alpha1"
@@ -64,6 +51,7 @@ func init() {
 //go:embed *.yaml
 var rawCustomResourceDefinitions embed.FS
 
+const crdName = "cowboys.wildwest.dev"
 const testNamespace = "cluster-controller-test"
 const clusterName = "us-east1"
 const sourceClusterName, sinkClusterName = "source", "sink"
@@ -173,7 +161,7 @@ func TestClusterController(t *testing.T) {
 				return
 			}
 			t.Log("Installing test CRDs...")
-			if err := installCrd(ctx, servers); err != nil {
+			if err := framework.InstallCrd(ctx, metav1.GroupKind{Group: wildwest.GroupName, Kind: "cowboys"}, servers, rawCustomResourceDefinitions); err != nil {
 				t.Error(err)
 				return
 			}
@@ -181,14 +169,14 @@ func TestClusterController(t *testing.T) {
 			start = time.Now()
 			source, sink := servers[sourceClusterName], servers[sinkClusterName]
 			t.Log("Installing sink cluster...")
-			if err := installCluster(t, ctx, source, sink); err != nil {
+			if err := framework.InstallCluster(t, ctx, source, sink, "clusters.cluster.example.dev", clusterName); err != nil {
 				t.Error(err)
 				return
 			}
 			t.Logf("Installed sink cluster after %s", time.Since(start))
 			start = time.Now()
 			t.Log("Setting up clients for test...")
-			if err := installNamespace(ctx, source); err != nil {
+			if err := framework.InstallNamespace(ctx, source, crdName, testNamespace); err != nil {
 				t.Error(err)
 				return
 			}
@@ -199,7 +187,7 @@ func TestClusterController(t *testing.T) {
 					t.Error(err)
 					return
 				}
-				clusterName, err := detectClusterName(cfg, ctx, "cowboys.wildwest.dev")
+				clusterName, err := framework.DetectClusterName(cfg, ctx, crdName)
 				if err != nil {
 					t.Errorf("failed to detect cluster name: %v", err)
 					return
@@ -242,156 +230,6 @@ func TestClusterController(t *testing.T) {
 			},
 		)
 	}
-}
-
-func installNamespace(ctx context.Context, server framework.RunningServer) error {
-	cfg, err := server.Config()
-	if err != nil {
-		return err
-	}
-	clusterName, err := detectClusterName(cfg, ctx, "cowboys.wildwest.dev")
-	if err != nil {
-		return fmt.Errorf("failed to detect cluster name: %w", err)
-	}
-	clients, err := kubernetesclientset.NewClusterForConfig(cfg)
-	if err != nil {
-		return fmt.Errorf("failed to construct client for server: %w", err)
-	}
-	client := clients.Cluster(clusterName)
-	_, err = client.CoreV1().Namespaces().Create(ctx, &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{Name: testNamespace},
-	}, metav1.CreateOptions{})
-	return err
-}
-
-func installCrd(ctx context.Context, servers map[string]framework.RunningServer) error {
-	wg := sync.WaitGroup{}
-	bootstrapErrChan := make(chan error, len(servers))
-	for _, server := range servers {
-		wg.Add(1)
-		go func(server framework.RunningServer) {
-			defer wg.Done()
-			cfg, err := server.Config()
-			if err != nil {
-				bootstrapErrChan <- err
-				return
-			}
-			crdClient, err := apiextensionsv1client.NewForConfig(cfg)
-			if err != nil {
-				bootstrapErrChan <- fmt.Errorf("failed to construct client for server: %w", err)
-				return
-			}
-			bootstrapErrChan <- config.BootstrapCustomResourceDefinitionFromFS(ctx, crdClient.CustomResourceDefinitions(), metav1.GroupKind{
-				Group: wildwest.GroupName,
-				Kind:  "cowboys",
-			}, rawCustomResourceDefinitions)
-		}(server)
-	}
-	wg.Wait()
-	close(bootstrapErrChan)
-	var bootstrapErrors []error
-	for err := range bootstrapErrChan {
-		bootstrapErrors = append(bootstrapErrors, err)
-	}
-	if err := kerrors.NewAggregate(bootstrapErrors); err != nil {
-		return fmt.Errorf("could not bootstrap CRDs: %w", err)
-	}
-	return nil
-}
-
-func installCluster(t framework.TestingTInterface, ctx context.Context, source, sink framework.RunningServer) error {
-	sourceCfg, err := source.Config()
-	if err != nil {
-		return fmt.Errorf("failed to get source config: %w", err)
-	}
-	rawSinkCfg, err := sink.RawConfig()
-	if err != nil {
-		return fmt.Errorf("failed to get sink config: %w", err)
-	}
-	sourceClusterName, err := detectClusterName(sourceCfg, ctx, "clusters.cluster.example.dev")
-	if err != nil {
-		return fmt.Errorf("failed to detect cluster name: %w", err)
-	}
-	sourceKcpClients, err := kcpclientset.NewClusterForConfig(sourceCfg)
-	if err != nil {
-		return fmt.Errorf("failed to construct client for server: %w", err)
-	}
-	rawSinkCfgBytes, err := clientcmd.Write(rawSinkCfg)
-	if err != nil {
-		return fmt.Errorf("failed to serialize sink config: %w", err)
-	}
-	sourceKcpClient := sourceKcpClients.Cluster(sourceClusterName)
-	cluster, err := sourceKcpClient.ClusterV1alpha1().Clusters().Create(ctx, &clusterv1alpha1.Cluster{
-		ObjectMeta: metav1.ObjectMeta{Name: clusterName},
-		Spec:       clusterv1alpha1.ClusterSpec{KubeConfig: string(rawSinkCfgBytes)},
-	}, metav1.CreateOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to create cluster on source kcp: %w", err)
-	}
-	defer source.Artifact(t, func() (runtime.Object, error) {
-		return sourceKcpClient.ClusterV1alpha1().Clusters().Get(ctx, cluster.Name, metav1.GetOptions{})
-	})
-	waitCtx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer func() {
-		cancel()
-	}()
-	watcher, err := sourceKcpClient.ClusterV1alpha1().Clusters().Watch(ctx, metav1.ListOptions{
-		FieldSelector: fields.OneTermEqualSelector("metadata.name", cluster.Name).String(),
-	})
-	if err != nil {
-		return fmt.Errorf("failed to watch cluster in source kcp: %w", err)
-	}
-	for {
-		select {
-		case <-waitCtx.Done():
-			return fmt.Errorf("failed to wait for cluster in source kcp to be ready: %w", waitCtx.Err())
-		case event := <-watcher.ResultChan():
-			switch event.Type {
-			case watch.Added, watch.Bookmark:
-				continue
-			case watch.Modified:
-				updated, ok := event.Object.(*clusterv1alpha1.Cluster)
-				if !ok {
-					continue
-				}
-				var ready bool
-				for _, condition := range updated.Status.Conditions {
-					if condition.Type == clusterv1alpha1.ClusterConditionReady && condition.Status == corev1.ConditionTrue {
-						ready = true
-						break
-					}
-				}
-				if ready {
-					return nil
-				}
-			case watch.Deleted:
-				return fmt.Errorf("cluster %s was deleted before being ready", cluster.Name)
-			case watch.Error:
-				return fmt.Errorf("encountered error while watching cluster %s: %#v", cluster.Name, event.Object)
-			}
-		}
-	}
-}
-
-// TODO: we need to undo the prefixing and get normal sharding behavior in soon ... ?
-func detectClusterName(cfg *rest.Config, ctx context.Context, crdName string) (string, error) {
-	crdClient, err := apiextensionsclientset.NewClusterForConfig(cfg)
-	if err != nil {
-		return "", fmt.Errorf("failed to construct client for server: %w", err)
-	}
-	crds, err := crdClient.Cluster("*").ApiextensionsV1().CustomResourceDefinitions().List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return "", fmt.Errorf("failed to list crds: %w", err)
-	}
-	if len(crds.Items) == 0 {
-		return "", errors.New("found no crds, cannot detect cluster name")
-	}
-	for _, crd := range crds.Items {
-		if crd.ObjectMeta.Name == crdName {
-			return crd.ObjectMeta.ClusterName, nil
-		}
-	}
-	return "", errors.New("detected no admin cluster")
 }
 
 // RegisterCowboyExpectation registers an expectation about the future state of the seed.
